working code

+++ src/controllers/userController.ts
@@ -1,4 +1,4 @@
-import { User } from '../models/index.js';
+import { User, Thought } from '../models/index.js';
 import { Request, Response } from 'express';
 
 
@@ -16,7 +16,7 @@ import { Request, Response } from 'express';
   export const getSingleUser = async (req: Request, res: Response) => {
     try {
       const user = await User.findOne({ _id: req.params.userId })
-        .select('-__v');
+        .select('-__v').populate('thoughts').populate('friends');
 
       if (!user) {
         return res.status(404).json({ message: 'No user with that ID' });
@@ -40,14 +40,35 @@ import { Request, Response } from 'express';
     }
   }
 
-  // Delete a user and associated apps
-  export const deleteUser = async (req: Request, res: Response) => {
+  export const updateUser = async (req: Request, res: Response) => {
     try {
-      const user = await User.findOneAndDelete({ _id: req.params.userId });
+      const user = await User.findOneAndUpdate({ _id: req.params.userId }, { $set: req.body }, { new: true, runValidators: true })
+        .select('-__v');
 
       if (!user) {
         return res.status(404).json({ message: 'No user with that ID' });
       }
+
+      res.json(user);
+      return;
+    } catch (err) {
+      res.status(500).json(err);
+      return;
+    }
+  }
+
+  // Delete a user and associated apps
+  export const deleteUser = async (req: Request, res: Response) => {
+    try {
+      const user = await User.findOne({ _id: req.params.userId });
+
+      if (!user) {
+        return res.status(404).json({ message: 'No user with that ID' });
+      }
+
+      await Thought.deleteMany({ _id: { $in: user.thoughts } });
+
+      await User.findOneAndDelete({ _id: req.params.userId });
 
       res.json({ message: 'User deleted!' })
       return;
@@ -57,3 +78,36 @@ import { Request, Response } from 'express';
     }
   }
 
+export const addFriend = async (req: Request, res: Response) => {
+  try {
+    const user = await User.findOneAndUpdate({ _id: req.params.userId }, 
+      { $addToSet: { friends: req.params.friendId } }, { new: true, runValidators: true })
+    .select('-__v');
+    if (!user) {
+      return res.status(404).json({ message: 'No user with that ID' });
+    }
+    res.json({ message: 'Friend added!' })
+    return;
+
+  } catch (err) {
+    res.status(500).json(err);
+    return;
+  }
+}
+
+export const unFriend = async (req: Request, res: Response) => {
+  try {
+    const user = await User.findOneAndUpdate({ _id: req.params.userId }, 
+      { $pull: { friends: req.params.friendId } }, { new: true, runValidators: true })
+    .select('-__v');
+    if (!user) {
+      return res.status(404).json({ message: 'No user with that ID' });
+    }
+    res.json({ message: 'Friend removed!' })
+    return;
+
+  } catch (err) {
+    res.status(500).json(err);
+    return;
+  }
+}
